+++ src/main/java/atomicedit/backend/blockprovider/ActionForBlock.java
@@ -7,5 +7,5 @@ package atomicedit.backend.blockprovider;
  * @author Justin Bonner
  */
 public interface ActionForBlock {
-    public void doAction(int x, int y, int z, short block);
+    public void doAction(int x, int y, int z, int block);
 }
